+++ src/main/java/sun/net/www/protocol/css/Handler.kt
@@ -1,5 +1,6 @@
 package sun.net.www.protocol.css
 
+import tornadofx.FX
 import tornadofx.Stylesheet
 import java.io.InputStream
 import java.net.URL
@@ -18,7 +19,9 @@ class Handler : URLStreamHandler() {
         override fun getInputStream(): InputStream {
             if (url.port == 64) return Base64.getDecoder().decode(url.host).inputStream()
             val stylesheet = Class.forName(url.host).newInstance() as Stylesheet
-            return stylesheet.render().byteInputStream(StandardCharsets.UTF_8)
+            val rendered = stylesheet.render()
+            if (FX.dumpStylesheets) println(rendered)
+            return rendered.byteInputStream(StandardCharsets.UTF_8)
         }
     }
 }